Improve func_8000A43C_B03C score

+++ src.us/core/AD60.c
@@ -120,7 +120,7 @@ void func_8000A3DC_AFDC(s16 arg0, s8 *arg1) {
 #ifdef NON_MATCHING
 s16 func_8000A43C_B03C(s8 *arg0) {
 	s8 var_v1;
-	u8 temp_v0;
+	s32 temp_v0;
 	s32 var_a0;
 
 	var_v1 = *arg0;
